refactor: Discord webhook 인터페이스/Config 제거, 서비스 단일 클래스로 통합 (#211)
